requirements.txt
streamlit
fpdf2
openai
plotly
pandas
numpy
requests